updates optic flow

+++ Go!/opticFlow.h
@@ -6,7 +6,7 @@ class OpticFlow :
 public:
 	OpticFlow();
 	virtual ~OpticFlow();
-	void createOpticFlow(osg::Array *points, const int mode = 0, const double size = 0, const double segments = 0);
+	void createOpticFlow(osg::Array *points, const int mode = 0, const double size = 0, const int segments = 0);
 	inline void setPointsColorArray(osg::ref_ptr<osg::Vec4Array> color) { _pointsColorArray = color.release(); };
 
 	inline void setFrameCounts(const unsigned &ref) { _frameCounts = ref; };
@@ -14,7 +14,7 @@ public:
 
 private:
 	void createGLPOINTS(osg::Vec3Array *p);
-	void createPolygon(osg::Vec3Array *p, const double radius, const double segments);
+	void createPolygon(osg::Vec3Array *p, const double radius, const int segments);
 	void createBalls(osg::Vec3Array *p, const double radius);
 	osg::ref_ptr<osg::Vec4Array> _pointsColorArray;
 	unsigned _frameCounts;
